Memory addresses are 16-bit now

+++ Assets/Scripts/Virtual CPU/Memory.cs
@@ -10,12 +10,12 @@ namespace NineEightOhThree.VirtualCPU
         public byte[] Data => data;
         public int size;
 
-        public HashSet<byte> writeProtectedAddresses;
+        public HashSet<ushort> writeProtectedAddresses;
 
         private void Awake()
         {
             Clear();
-            writeProtectedAddresses = new HashSet<byte>();
+            writeProtectedAddresses = new HashSet<ushort>();
         }
 
         public void Clear()
@@ -23,17 +23,17 @@ namespace NineEightOhThree.VirtualCPU
             data = new byte[size];
         }
 
-        public byte Read(byte address)
+        public byte Read(ushort address)
         {
             return data[address];
         }
 
-        public byte Read(byte address, byte offset)
+        public byte Read(ushort address, byte offset)
         {
-            return Read((byte)(address + offset));
+            return Read((ushort)(address + offset));
         }
 
-        public bool Write(byte address, byte value)
+        public bool Write(ushort address, byte value)
         {
             if (writeProtectedAddresses.Contains(address))
                 return false;
@@ -41,25 +41,25 @@ namespace NineEightOhThree.VirtualCPU
             return true;
         }
 
-        public bool Write(byte address, byte offset, byte value)
+        public bool Write(ushort address, byte offset, byte value)
         {
-            return Write((byte)(address + offset), value);
+            return Write((ushort)(address + offset), value);
         }
 
-        public byte[] ReadBlock(byte address, int count)
+        public byte[] ReadBlock(ushort address, int count)
         {
             return data[address..(address + count)];
         }
 
-        public void ProtectWrites(params byte[] addresses)
+        public void ProtectWrites(params ushort[] addresses)
         {
-            foreach (byte address in addresses)
+            foreach (ushort address in addresses)
                 writeProtectedAddresses.Add(address);
         }
 
-        public void UnprotectWrites(params byte[] addresses)
+        public void UnprotectWrites(params ushort[] addresses)
         {
-            foreach (byte address in addresses)
+            foreach (ushort address in addresses)
                 writeProtectedAddresses.Remove(address);
         }
     }
